+++ pkg/corerp/frontend/controller/environments/v20250801preview/test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package environments_v20250801preview
+package v20250801preview
 
 import (
 	"encoding/json"
