+++ pkg/corerp/frontend/controller/environments/v20250801preview/testdata/environmentresource_input.json
